requirements.txt
# ── Web Framework ─────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0

# ── Async Database ─────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Data Validation ────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Authentication & Security ──────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Caching & Rate Limiting ────────────────────────────────────────
redis[hiredis]==5.0.4
fastapi-limiter==0.1.6

# ── Task Queue ─────────────────────────────────────────────────────
celery[rabbitmq]==5.4.0
flower==2.0.1

# ── Machine Learning ───────────────────────────────────────────────
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# ── HTTP Client ────────────────────────────────────────────────────
httpx==0.27.0
aiohttp==3.9.5

# ── Utilities ──────────────────────────────────────────────────────
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0
pytz==2024.1

# ── Testing ────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0
httpx==0.27.0
factory-boy==3.3.0
faker==25.2.0
locust==2.28.0
